perf: rename constants to globals
